scripts/followup/requirements.txt
astroid==2.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.6.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
